When generating the C header, assign predictable values to enum keys
When tzplatform-tool is used to generate C header file, it puts enum keys
in the input order. This tends to be problematic when new keys are added
in the middle of the input. Newly generated header shifts old enum keys
and gives them different values than before. This breaks clients compiled
against the old header, using different values for enum keys.
This patch attempts to mitigate it. Enum keys now get explicit value, based
on hash of the key name. Adding a new keys won't change values assigned to
the old ones.

Change-Id: I87803419684f7e0eeda90bf2df757587468f9dab
Signed-off-by: Rafal Krypa <r.krypa@samsung.com>