gcc 5.2.0: fix linking issues caused by inline keyword
Under gcc 5.2.0, using the inline keyword on function declarations in
a header file without also providing the implementation leads first to
warnings, then during linking of a library using the functions to an
error:

 In file included from shared-api.c:48:0:
 init.h:27:13: warning: inline function 'initialize' declared but never defined
  inline void initialize(struct tzplatform_context *context);
 ...
 init.c:426:29: warning: 'metafilepath' is static but used in inline function 'initialize' which is not static
                  keyname(i), metafilepath);
 ...
 ./.libs/libtzplatform-config-2.0.so: undefined reference to `initialize'
 ./.libs/libtzplatform-config-2.0.so: undefined reference to `hashid'
 ./.libs/libtzplatform-config-2.0.so: undefined reference to `get_uid'
 collect2: error: ld returned 1 exit status
 Makefile:573: recipe for target 'tzplatform-get' failed

This kind of inlining cannot have worked before; probably older gcc
simply ignored the keyword (untested). Therefore removing the keyword
is the quickest way to get the code to compile again as before. A more
intrusive change would be needed if inlining is really important.

Change-Id: Ic88350071ddb2df6bf571ed60b3cf6d443aba742
Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>